+++ qpi-driver/qpi_driver/cli.py
@@ -153,7 +153,7 @@ if typer.IS_TYPER_INSTALLED:
         try:
             return importlib.metadata.version("qpi-driver")
         except importlib.metadata.PackageNotFoundError:
-            return "0.0.32"
+            return "0.0.33"
 
     def _banner():
         """Renders the banner at the top of the CLI"""